Stats Page functionality as well as a timer.
The stats page now works and should be easy to update as time goes on. UI is eh but we can make the game look pretty later. I added a timer for the time played stat track, but it also functions in that we can use it for random events later. Also the stats page is updated every second, bound to the timer so it is pretty robust.